Split resolve errors page into layout and index components
