+++ frontend/src/features/resolve_errors/routes.tsx
@@ -1,5 +1,11 @@
 import { RouteObject } from "react-router";
 
-import { ResolveErrorsPage } from "./components/ResolveErrorsPage";
+import { ResolveErrorsIndexPage } from "./components/ResolveErrorsIndexPage";
+import { ResolveErrorsLayout } from "./components/ResolveErrorsLayout";
 
-export const resolveErrorsRoutes: RouteObject[] = [{ index: true, Component: ResolveErrorsPage }];
+export const resolveErrorsRoutes: RouteObject[] = [
+  {
+    Component: ResolveErrorsLayout,
+    children: [{ index: true, Component: ResolveErrorsIndexPage }],
+  },
+];
